refactor(llm): move openai tool_search into model_constructors
