+++ front/lib/model_constructors/sdk/openai_responses/converters/input/tool_search.ts
